Move models into new files and buff domain message models
